Fix minor mistakes in session2

+++ exercises/04_data_type_sizes/data_type_sizes.c
@@ -10,7 +10,7 @@ int main() {
            sizeof(char), CHAR_MIN, CHAR_MAX);
     printf("The size of int is %zu bytes; it ranges from %d to %d\n",
            sizeof(int), INT_MIN, INT_MAX);
-    printf("The size of uint32_t is %zu bytes\n", sizeof(int16_t));
+    printf("The size of uint32_t is %zu bytes\n", sizeof(uint32_t));
 
     printf("The largest double that can be represented is %6.1e\n", DBL_MAX);
     printf("The smallest positive double that can be represented is %6.1e\n", DBL_MIN);
